CECS 130 ASsigment
This is my favorite assignment that I completed in my CECS 130 programming class. I took this assignment to the next level by adding more fuctionality than listed in the problem statement. The problem statement only wanted a program that would ask a user to input the amount of dice they wanted to be rolled bewtween 1 and six then print out the resulting rolls.